requirements.txt
Flask==3.1.3
gunicorn==25.3.0
psycopg2-binary==2.9.11